# backend/swagger.yml
openapi: 3.0.0
info:
  title: AWS-Test
  description: API Docs for AWS-Test server
  version: 0.0.1
servers:
  - url: https://18.222.161.136:5000
  - url: https:localhost:5000
  - url: /
paths:
  /test:
    get:
      summary: Returns a JSON containing a message.
      responses:
        200:
          description: "Success"
          schema:
            type: string
  /register:
    post:
      summary: Adds a new user into the database
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        200:
          description: "Success"
        400:
          description: "Bad Request"
  /login:
    post:
      summary: Login with user credentials
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        200:
          description: "Success"
        400:
          description: "Bad Request"
  /profile:
    post:
      summary: Test token bullshit
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                accessToken:
                  type: string
      responses:
        200:
          description: "Success"
        400:
          description: "Bad Request"
  /addWorkspace:
    post:
      summary: Add Workspace
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                accessToken:
                  type: string
                username:
                  type: string
                workspaceName:
                  type: string
      responses:
        200:
          description: "Success"
        400:
          descripton: "Bad Request"
  /getWorkspaces:
    post:
      summary: Get All Workspaces
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                accessToken:
                  type: string
      responses:
        200:
          description: "Retrieved Workspaces"
        400:
          description: "Bad Request"
  /addUserToWorkspace:
    post:
      summary: Add User to Workspace
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                accessToken:
                  type: string
                username:
                  type: string
                workspaceName:
                  type: string
                workspaceOwner:
                  type: string
      responses:
        200:
          description: "Added user to workspace"            
        400:
          description: "Bad request"